Moved to versioned API, removed shapely dependency
